zktime_windows_app/requirements.txt
# ZKTeco Windows Application Requirements
# Core dependencies
requests>=2.28.0
schedule>=1.2.0

# GUI and System Tray
pystray>=0.19.4
Pillow>=9.0.0

# ZKTeco device communication
pyzk>=0.9.0

# Additional utilities
python-dateutil>=2.8.2 